+++ packages/components/scripts/generate_icons.js
@@ -198,7 +198,7 @@ import * as Tabler from './TablerIcons';
 import { IconProps } from './types';
 
 const withIcon = (Component: React.FC, props: IconProps) => {
-  const styles = useStyleConfig('Icon') as Record<
+  const styles = useStyleConfig('Icon', { size: props?.size }) as Record<
     string,
     RecursiveCSSObject<CSSWithMultiValues>
   >;
@@ -223,10 +223,17 @@ const storiesContent = `${doNotEditHeader}
 import React from 'react';
 import { Meta } from '@storybook/react';
 import * as Icons from './Icons';
+import { iconSizes } from './types';
 import { Box, Flex, Grid } from 'components';
 
 export default {
   title: 'Components/Primitives/Icons',
+  argTypes: {
+    size: {
+      options: iconSizes,
+      control: { type: 'select' },
+    },
+  },
 } as Meta;
 
 const Template = (args) => (
@@ -246,7 +253,6 @@ const Template = (args) => (
 export const All = Template.bind({});
 All.args = {
   color: 'primary-default',
-  boxSize: '1.5rem',
 };
 `;