+++ packages/components/src/components/primitives/Breadcrumb/types.ts
@@ -1,0 +1,18 @@
+export {
+  BreadcrumbProps,
+  BreadcrumbItemProps,
+  BreadcrumbLinkProps,
+  BreadcrumbSeparatorProps,
+} from '@chakra-ui/react';
+import {
+  BreadcrumbItemProps,
+  BreadcrumbLinkProps,
+  BreadcrumbSeparatorProps,
+} from '@chakra-ui/react';
+import { FC } from 'react';
+
+export interface BreadcrumbStaticMembers {
+  Item: FC<BreadcrumbItemProps>;
+  Link: FC<BreadcrumbLinkProps>;
+  Separator: FC<BreadcrumbSeparatorProps>;
+}