+++ packages/components/src/theme/ComposerTheme/Components/Icon.ts
@@ -1,7 +1,21 @@
 import { ComponentStyleConfig } from '@chakra-ui/theme';
 
 export const Icon: ComponentStyleConfig = {
-  baseStyle: {
-    boxSize: '1.5rem',
+  sizes: {
+    xs: {
+      boxSize: '0.75rem',
+    },
+    s: {
+      boxSize: '1rem',
+    },
+    m: {
+      boxSize: '1.25rem',
+    },
+    l: {
+      boxSize: '1.5rem',
+    },
+  },
+  defaultProps: {
+    size: 'm',
   },
 };