+++ packages/components/src/theme/ComposerTheme/Components/index.ts
@@ -1,5 +1,7 @@
+import { Accordion } from './Accordion';
 import { Avatar, AvatarBadge } from './Avatar';
 import { Badge } from './Badge';
+import { Breadcrumb } from './Breadcrumb';
 import { Alert } from './Alert';
 import { Button } from './Button';
 import { Card } from './Card';
@@ -9,17 +11,24 @@ import { Divider } from './Divider';
 import { Icon } from './Icon';
 import { Input } from './Input';
 import { Select } from './Select';
+import { Link } from './Link';
+import { Progress } from './Progress';
 import { Slider } from './Slider';
 import { Spinner } from './Spinner';
+import { Switch } from './Switch';
 import { Tabs } from './Tabs';
+import { Tag } from './Tag';
 import { Text } from './Text';
 import { TextPairing } from './TextPairing';
+import { Radio } from './Radio';
 
 export const components = {
+  Accordion,
   Alert,
   Avatar,
   AvatarBadge,
   Badge,
+  Breadcrumb,
   Button,
   Card,
   CircularProgress,
@@ -27,10 +36,15 @@ export const components = {
   Divider,
   Icon,
   Input,
+  Link,
+  Progress,
   Select,
+  Radio,
   Slider,
   Spinner,
+  Switch,
   Tabs,
+  Tag,
   Text,
   TextPairing,
 };