+++ packages/components/src/theme/ComposerTheme/styles.ts
@@ -7,7 +7,7 @@ const baseTextStyle = {
 const linkTextStyle = {
   ...baseTextStyle,
   color: 'text-link-accent-default',
-  textDecorationLine: 'underline',
+  textDecoration: 'underline',
 };
 
 export const textStyles = {